requirements.txt
fastapi
uvicorn
python-multipart
jinja2
langgraph
langchain
langchain-openai
pypdf
pandas
openpyxl
python-docx
google-generativeai
qdrant-client
sentence-transformers
python-dotenv
torch
matplotlib
reportlab
langchain-experimental
sympy
pillow

